Added Eratosthenes sieve + tests, changed the structure of integers math (added new Object for hardcore math generators)
