+++ README.md
@@ -71,7 +71,11 @@ Would be great, if you'd contribute, in case that presented algorithms are not a
 10) Fibonacci numbers
 11) Random ints
 12) Catalan numbers
-13) Fermat numbers
+
+### Additional math generators:
+[`integerOperations.IntegerGeneratorsMath`](src/main/scala/integerOperations/IntegerGeneratorsMath.scala)
+1) Fermat numbers
+2) Eratosthenes primes sieve O(log(log(n)))
 
 ### Char operations:
 [`charOperations.CharProperties`](src/main/scala/charOperations/CharProperties.scala)
