+++ src/main/scala/Main.scala
@@ -3,6 +3,7 @@ import listOperations.ListProperties._
 import scala.concurrent.ExecutionContext.Implicits.global
 import charOperations.CharProperties._
 import integerOperations.IntegerProperties._
+import integerOperations.IntegerGeneratorsMath._
 import integerOperations.IntegerMath._
 import integerOperations.IntegerGenerators._
 import doubleOperations.DoubleProperties._
@@ -16,6 +17,6 @@ import scala.concurrent.Future
 
 object Main extends App {
 
-  print( 17.isPrimeFermatStrict() )
+  print( 17.generateEratosthenesPrimes )
 
 }
