+++ src/test/scala/unitTests/IntegerTests.scala
@@ -135,14 +135,6 @@ class IntegerTests extends FunSuite {
     0.binaryPower(5) shouldBe 0
   }
 
-  test("Testing of generateFermatNumbers") {
-    258.generateFermatNumbers shouldBe List(3, 5, 17, 257)
-    assertThrows[InputException] {
-      (-1).generateFermatNumbers
-    }
-    3.generateFermatNumbers shouldBe List(3)
-  }
-
   test("Testing of isPrimeFermatStrict") {
     2.isPrimeFermatStrict() shouldBe true
     561.isPrimeFermatStrict() shouldBe false
